rename run_program_op to run_program, its not a op
